Fix "generated CSS" panel when starting in preview mode
